Fix STOR-901 - prevent extending of expired sessions
Expired session is marked with BeginTime = NULL in the DB that not happends when IIS kills AccessContol based on Idle site timeout.
If session will expire when AccessControl is not working (active) expired session (EndDate after current date) still can be extended because we didn't check for EndTime expiration.

Such sessions are not marked as expired even when AccessControl is restarted - I will fix this in Orion.