Fix: Enable TLS 1.1 and 1.2 for Server initiated HTTPS calls
When TLS 1.0 is disable for the server (that is common practice now because of security issues) main site (JX) cannot connect to FileStore (and other public services - in status call) because in .NET 4.5.2 only SSL3 and TLS 1.0 are enabled by default. This fix will keep support/use of these old protocols but also enable the modern ones.